requirements.txt
fastapi==0.95.2
fastapi-cache2==0.2.1
uvicorn==0.15.0
numpy==1.23.4
aiocache==0.12.1
aiohttp==3.8.4
psutil==5.9.5
requests-mock==1.10.0
seaborn
matplotlib
python-dotenv
requests
pytest
pytest-cov